requirements.txt
google-generativeai==0.4.1
python-dotenv==1.0.1
feedparser==6.0.11
requests==2.32.3
trafilatura==1.9.0
tenacity==8.5.0